+++ data_pump/var_declarations.py
@@ -1,0 +1,24 @@
+labels = dict()
+eperson_id = dict()
+email2epersonId = dict()
+userRegistration_id = dict()
+group_id = dict()
+metadata_schema_id = dict()
+metadata_field_id = dict()
+community_id = dict()
+community2logo = dict()
+collection_id = dict()
+collection2logo = dict()
+item_id = dict()
+workspaceitem_id = dict()
+workflowitem_id = dict()
+metadatavalue = dict()
+handle = dict()
+bitstreamformat_id = dict()
+primaryBitstream = dict()
+bitstream2bundle = dict()
+bundle_id = dict()
+bitstream_id = dict()
+statistics = dict()
+unknown_format_id = None
+imported_handle = 0
